+++ src/main/java/spring/backend/recommendation/application/GetRecommendationsFromClovaService.java
@@ -7,8 +7,8 @@ import spring.backend.activity.domain.value.Keyword;
 import spring.backend.activity.domain.value.Keyword.Category;
 import spring.backend.activity.exception.ActivityErrorCode;
 import spring.backend.core.converter.ImageConverter;
-import spring.backend.recommendation.dto.request.AIRecommendationRequest;
-import spring.backend.recommendation.dto.response.ClovaRecommendationResponse;
+import spring.backend.recommendation.presentation.dto.request.AIRecommendationRequest;
+import spring.backend.recommendation.presentation.dto.response.ClovaRecommendationResponse;
 import spring.backend.recommendation.infrastructure.clova.dto.response.ClovaResponse;
 import spring.backend.recommendation.infrastructure.clova.exception.ClovaErrorCode;
 import spring.backend.recommendation.infrastructure.map.kakao.dto.response.KakaoMapResponse;
