+++ src/main/java/spring/backend/recommendation/application/GetRecommendationsFromOpenAIService.java
@@ -8,8 +8,8 @@ import spring.backend.activity.domain.value.Keyword;
 import spring.backend.activity.domain.value.Keyword.Category;
 import spring.backend.activity.domain.value.Type;
 import spring.backend.core.converter.ImageConverter;
-import spring.backend.recommendation.dto.request.AIRecommendationRequest;
-import spring.backend.recommendation.dto.response.OpenAIRecommendationResponse;
+import spring.backend.recommendation.presentation.dto.request.AIRecommendationRequest;
+import spring.backend.recommendation.presentation.dto.response.OpenAIRecommendationResponse;
 import spring.backend.recommendation.infrastructure.dto.Message;
 import spring.backend.recommendation.infrastructure.openai.dto.response.OpenAIResponse;
 import spring.backend.recommendation.infrastructure.openai.dto.response.OpenAIResponse.Choice;
