+++ src/main/java/spring/backend/recommendation/application/RecommendationProvider.java
@@ -1,6 +1,6 @@
 package spring.backend.recommendation.application;
 
-import spring.backend.recommendation.dto.request.AIRecommendationRequest;
+import spring.backend.recommendation.presentation.dto.request.AIRecommendationRequest;
 
 public interface RecommendationProvider<T> {
     T getRecommendations(AIRecommendationRequest aiRecommendationRequest);
